requirements.txt
absl-py==0.11.0
alive-progress==1.6.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1600699180754/work
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
h5py==3.1.0
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603376012865/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1598113097404/work
pandas @ file:///tmp/build/80754af9/pandas_1602088128026/work
Pillow @ file:///tmp/build/80754af9/pillow_1603822253009/work
protobuf==3.13.0
py-polar-codes==1.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing==2.4.7
PySimpleGUI==4.30.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
torch==1.2.0
torchvision==0.4.0a0+6b959ee
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
